backend/requirementsFilterRoberta.txt
numpy
streamlit==0.84.1
Pillow
jax[cpu]
flax
torch
torchvision
transformers
huggingface_hub
googletrans==4.0.0-rc1
protobuf==3.20
scipy.special